add montant link only visible for bureau

+++ resources/views/tresorerie/livret.blade.php
@@ -32,15 +32,17 @@
                             <td>
                                 {{$livret->created_at}}
                             </td>
-                            <td>
-                                <a href="#editlivret{{$livret->id}}"
-                                   class="btn btn-primary btn-xs"
-                                   data-toggle="modal"><i class="fa fa-pencil"></i></a>
-                                <a href="{{route('delete.montant', $livret->id)}}"
-                                   data-method="delete"
-                                   data-confirm="Souhaitez-vous réellement supprimer ce montant ?"
-                                   class="btn btn-danger btn-xs"><i class="fa fa-trash-o "></i></a>
-                            </td>
+                            @if(Auth::user()->statut_id == 2)
+                                <td>
+                                    <a href="#editlivret{{$livret->id}}"
+                                       class="btn btn-primary btn-xs"
+                                       data-toggle="modal"><i class="fa fa-pencil"></i></a>
+                                    <a href="{{route('delete.montant', $livret->id)}}"
+                                       data-method="delete"
+                                       data-confirm="Souhaitez-vous réellement supprimer ce montant ?"
+                                       class="btn btn-danger btn-xs"><i class="fa fa-trash-o "></i></a>
+                                </td>
+                            @endif
                         </tr>
                         @include('tresorerie.edit')
                     @endforeach
